# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ERP System API — PR/PO/Store Management with full Approval workflow",
        "title": "ERP API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ERP Support",
            "email": "support@erp.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pr": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase Request"
                ],
                "summary": "List purchase requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase Request"
                ],
                "summary": "Create purchase request",
                "parameters": [
                    {
                        "description": "PR data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePRRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/pr/next-number": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the next available PR number for the current month (Buddhist year format)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase Request"
                ],
                "summary": "Get next PR number",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/pr/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase Request"
                ],
                "summary": "Get PR detail with lines and attachments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PR ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
        "/pr/{id}/approve": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase Request"
                ],
                "summary": "Approve a purchase request",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PR ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase Request"
                ],
                "summary": "Approve or reject PR (Senior Team Project)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PR ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "action",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ApprovalActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
        "/pr/{id}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase Request"
                ],
                "summary": "Get PR status history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PR ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/fiber.Map"
                            }
                        }
                    }
                }
            }
        },
...
        "/pr/{id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase Request"
                ],
                "summary": "Submit PR for approval",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PR ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/fiber.Map"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "fiber.Map": {
            "type": "object",
            "additionalProperties": true
        },
        "models.AddAttachmentRequest": {
            "type": "object",
            "required": [
                "file_name",
                "file_path"
            ],
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "type": "string"
                }
            }
        },
        "models.ApprovalActionRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "type": "string",
                    "enum": [
                        "APPROVE",
                        "REJECT",
                        "RETURN",
                        "COMMENT"
                    ]
                },
                "comments": {
                    "type": "string"
                }
            }
        },
...
        "models.CreatePRLine": {
            "type": "object",
            "required": [
                "mat_code",
                "qty_requested"
            ],
            "properties": {
                "line_no": {
                    "type": "integer"
                },
                "mat_code": {
                    "type": "string"
                },
                "qty_requested": {
                    "type": "number"
                }
            }
        },
        "models.CreatePRRequest": {
            "type": "object",
            "required": [
                "created_by",
                "lines",
                "location_code",
                "pr_date",
                "pr_no",
                "requested_by"
            ],
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AddAttachmentRequest"
                    }
                },
                "created_by": {
                    "type": "integer"
                },
                "lines": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.CreatePRLine"
                    }
                },
                "location_code": {
                    "type": "string"
                },
                "pr_date": {
                    "type": "string"
                },
                "pr_no": {
                    "type": "string"
                },
                "project_code": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "requested_by": {
                    "type": "integer"
                },
                "required_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}